refactor: update testnet naming
