+++ .env.dev
@@ -37,4 +37,4 @@ ARBITRUM_RPC = https://go.getblock.io/1244cf394ef143ae9b4eac44899ec6f4
 
 L1BitcoinDepositor = 0xD9B523fb879C63b00ef14e48C98f4e3398d3BA2D
 L2BitcoinDepositor = 0xB2fEC598a9374078Bb639f3d70555fc4389b7a78
-TBTCVaultSepolia = 0xB5679dE944A79732A75CE556191DF11F489448d5
+TBTCVault = 0xB5679dE944A79732A75CE556191DF11F489448d5
